# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: R&S Admissao API
  description: |
    API do sistema de Recrutamento e Selecao (R&S) para agendamento de aulas teste.

    Este sistema permite que candidatos agendem aulas teste em unidades educacionais,
    enquanto a equipe de recrutamento gerencia vagas, configuracoes de horarios e
    bloqueios de agenda.

    ## Autenticacao

    A API utiliza dois mecanismos de autenticacao:

    - **JWT Bearer Token**: Para usuarios autenticados (equipe de recrutamento)
    - **Acesso Publico**: Alguns endpoints sao publicos para candidatos

    Endpoints marcados com `optionalAuth` podem ser acessados com ou sem autenticacao,
    com comportamento diferenciado baseado no contexto.

    ## Ambientes

    | Ambiente | URL Base |
    |----------|----------|
    | Producao | https://api.rsadmissao.tomeducacao.com.br/api/v1 |
    | Homologacao | https://api-homolog.rsadmissao.tomeducacao.com.br/api/v1 |

    ## Versionamento

    A API segue versionamento semantico. A versao atual e `v1`.

    ## Rate Limiting

    A API implementa limitacao de taxa para prevenir abuso:

    | Endpoint | Limite | Janela |
    |----------|--------|--------|
    | `POST /auth/login` | 5 requisicoes | 1 minuto |
    | `POST /booking` (publico) | 10 requisicoes | 1 minuto |
    | `GET /availability` | 60 requisicoes | 1 minuto |
    | Demais endpoints | 100 requisicoes | 1 minuto |

    **Headers de resposta:**
    - `X-RateLimit-Limit`: Limite total de requisicoes
...

...

...

...

...
servers:
...
paths:
...
  /availability:
    get:
      tags:
        - availability
      summary: Consultar slots disponiveis
      description: |
        Retorna os slots disponiveis para agendamento de aulas teste.

        O comportamento varia conforme o contexto de autenticacao:

        | Aspecto | Publico (sem auth) | Privado (com auth) |
        |---------|--------------------|--------------------|
        | Range de datas | Menor entre d_rules e valid_from/valid_until | valid_from ate valid_until |
        | Sabado | Excluido | Incluido |
        | Status dos slots | Apenas `vago` | Todos (vago, ocupado, bloqueado) |
        | Config exposta | Apenas slot_size | Configuracao completa |

        A paginacao e feita por semana, usando domingo como separador.
      operationId: getAvailability
      security:
        - {}
        - bearerAuth: []
      parameters:
        - name: id_unidade
          in: query
          required: true
          description: ID da unidade para consultar disponibilidade
          schema:
            type: integer
            example: 123
        - name: page
          in: query
          required: false
          description: |
            Numero da pagina (semana) a ser retornada.
            Se omitido, retorna a semana que contem a data atual.
          schema:
            type: integer
            minimum: 1
            example: 1
      responses:
        '200':
          description: Slots disponiveis retornados com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvailabilityResponse'
              examples:
                publico:
                  summary: Resposta para acesso publico
                  value:
                    success: true
                    config:
                      slot_size: 50
                    pagination:
                      current_page: 1
                      total_pages: 4
                      week_start: "2025-01-06"
                      week_end: "2025-01-10"
                    slots:
                      - date: "2025-01-06"
                        day_of_week: "segunda-feira"
                        slot_start: "08:00:00"
                        slot_end: "08:50:00"
                        status: "vago"
                      - date: "2025-01-06"
                        day_of_week: "segunda-feira"
                        slot_start: "09:00:00"
                        slot_end: "09:50:00"
                        status: "vago"
                privado:
                  summary: Resposta para acesso autenticado
                  value:
                    success: true
                    config:
                      slot_size: 50
                      morning_start_at: "08:00:00"
                      morning_end_at: "12:00:00"
                      afternoon_start_at: "13:00:00"
                      afternoon_end_at: "17:00:00"
                      d_rule_start: 1
                      d_rule_end: 30
                      valid_from: "2025-01-01"
                      valid_until: "2025-03-31"
                    pagination:
                      current_page: 1
                      total_pages: 4
                      week_start: "2025-01-06"
                      week_end: "2025-01-11"
                    slots:
                      - date: "2025-01-06"
                        day_of_week: "segunda-feira"
                        slot_start: "08:00:00"
                        slot_end: "08:50:00"
                        status: "vago"
                      - date: "2025-01-06"
                        day_of_week: "segunda-feira"
                        slot_start: "09:00:00"
                        slot_end: "09:50:00"
                        status: "ocupado"
                        booking_id: 456
                        candidate_name: "Maria Silva"
        '400':
          description: Parametros invalidos ou periodo nao disponivel
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missingIdUnidade:
                  summary: id_unidade nao informado
                  value:
                    success: false
                    errors:
                      - msg: "id_unidade e obrigatorio"
                        param: "id_unidade"
                        location: "query"
                periodoIndisponivel:
                  summary: Periodo de agendamento nao disponivel
                  value:
                    success: false
                    error: "Periodo de agendamento nao disponivel"
        '404':
          description: Configuracao de horarios nao encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Configuracao de horarios nao encontrada"
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ============================================================================
  # BOOKING
  # ============================================================================
  /booking:
    get:
      tags:
        - booking
      summary: Listar agendamentos
      description: |
        Lista agendamentos com filtros opcionais por unidade e/ou CPF.

        Requer autenticacao com role `recrutamento` ou `admin`.
      operationId: listBookings
      security:
        - bearerAuth: []
      parameters:
        - name: id_unidade
          in: query
          required: false
          description: Filtrar por ID da unidade
          schema:
            type: integer
            example: 123
        - name: cpf
          in: query
          required: false
          description: Filtrar por CPF do candidato (somente digitos)
          schema:
            type: string
            pattern: '^\d{11}$'
            example: "12345678901"
        - name: status
          in: query
          required: false
          description: Filtrar por status do agendamento
          schema:
            $ref: '#/components/schemas/BookingStatus'
        - name: start_date
          in: query
          required: false
          description: Data inicial do periodo (formato YYYY-MM-DD)
          schema:
            type: string
            format: date
            example: "2025-01-01"
        - name: end_date
          in: query
          required: false
          description: Data final do periodo (formato YYYY-MM-DD)
          schema:
            type: string
            format: date
            example: "2025-01-31"
      responses:
        '200':
          description: Lista de agendamentos
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  bookings:
                    type: array
                    items:
                      $ref: '#/components/schemas/BookingListItem'
              example:
                success: true
                bookings:
                  - id_booking: 1
                    id_job_unidade: 10
                    id_application_gupy: "app_12345"
                    cpf: "12345678901"
                    candidate_name: "Maria Silva"
                    candidate_email: "maria@email.com"
                    start_at: "2025-01-15T09:00:00-03:00"
                    end_at: "2025-01-15T09:50:00-03:00"
                    status_booking: "agendado"
                    unidade_nome: "Escola Exemplo"
                    job_name: "Professor de Matematica"
                    created_at: "2025-01-10T14:30:00-03:00"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - booking
      summary: Criar agendamento
      description: |
        Cria um novo agendamento de aula teste.

        **Fluxo Publico (sem auth):**
        - Requer validacao do candidato via Gupy (middleware `validateApplication`)
        - Aplica d_rules para validar se o horario esta no periodo permitido
        - Cria eventos no Google Calendar para coordenador e candidato

        **Fluxo Privado (com auth):**
        - Permite agendamento manual pela equipe de recrutamento
        - Bypassa d_rules (pode agendar fora do periodo publico)
        - Marca o agendamento como manual no evento do calendario

        **Regras de Negocio:**
        - Candidato pode ter no maximo 1 agendamento ativo (status = 'agendado')
        - Candidato pode ter no maximo 1 aula teste realizada por unidade (status = 'compareceu')
        - Slot deve estar disponivel (nao ocupado ou bloqueado)
      operationId: createBooking
      security:
        - {}
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
            examples:
              publico:
                summary: Agendamento por candidato
                value:
                  id_job_unidade: 10
                  jobId: "job_gupy_123"
                  applicationId: "app_gupy_456"
                  start_at: "2025-01-15T09:00:00-03:00"
              privado:
                summary: Agendamento manual por recrutamento
                value:
                  id_job_unidade: 10
                  id_application_gupy: "app_gupy_456"
                  cpf: "12345678901"
                  candidate_name: "Maria Silva"
                  candidate_email: "maria@email.com"
                  start_at: "2025-01-15T09:00:00-03:00"
      responses:
        '201':
          description: Agendamento criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
              example:
                success: true
                message: "Agendamento criado com sucesso"
                booking:
                  id_booking: 1
                  id_job_unidade: 10
                  id_application_gupy: "app_gupy_456"
                  cpf: "12345678901"
                  start_at: "2025-01-15T09:00:00-03:00"
                  end_at: "2025-01-15T09:50:00-03:00"
                  status_booking: "agendado"
                  id_calendar_event: "cal_event_123"
                  id_calendar_event_candidato: "cal_event_456"
                  created_at: "2025-01-10T14:30:00-03:00"
        '400':
          description: Dados invalidos ou regra de negocio violada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                slotOcupado:
                  summary: Slot ja ocupado
                  value:
                    success: false
                    error: "Horario nao disponivel para agendamento"
                candidatoComAgendamento:
                  summary: Candidato ja possui agendamento ativo
                  value:
                    success: false
                    error: "Candidato ja possui agendamento ativo"
                candidatoJaRealizou:
                  summary: Candidato ja realizou aula teste na unidade
                  value:
                    success: false
                    error: "Candidato ja realizou aula teste nesta unidade"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Candidato nao elegivel
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Candidato nao esta na fase para agendamento de aula teste"
                current_step: "Entrevista"
        '404':
          description: Vaga ou candidatura nao encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Vaga nao encontrada"
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /schedule-block:
    get:
      tags:
        - schedule-block
      summary: Listar bloqueios de horario
      description: |
        Lista os bloqueios de horario configurados.

        Bloqueios impedem que candidatos agendem em determinados periodos
        (ex: ferias do coordenador, reunioes, etc).

        **Regra de prioridade:** Booking > Block. Agendamentos existentes
        nao sao afetados por bloqueios criados posteriormente.

        Requer autenticacao com role `recrutamento` ou `admin`.
      operationId: listScheduleBlocks
      security:
        - bearerAuth: []
      parameters:
        - name: id_unidade
          in: query
          required: false
          description: Filtrar por ID da unidade
          schema:
            type: integer
            example: 123
        - name: active_only
          in: query
          required: false
          description: Retornar apenas bloqueios ativos (data_fim >= hoje)
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Lista de bloqueios
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  blocks:
                    type: array
                    items:
                      $ref: '#/components/schemas/ScheduleBlock'
              example:
                success: true
                blocks:
                  - id_schedule_block: 1
                    id_unidade: 123
                    date_start: "2025-01-20"
                    date_end: "2025-01-24"
                    time_start: null
                    time_end: null
                    reason: "Ferias do coordenador"
                    created_at: "2025-01-10T10:00:00-03:00"
                  - id_schedule_block: 2
                    id_unidade: 123
                    date_start: "2025-02-01"
                    date_end: "2025-02-01"
                    time_start: "14:00:00"
                    time_end: "16:00:00"
                    reason: "Reuniao pedagogica"
                    created_at: "2025-01-15T09:00:00-03:00"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - schedule-block
      summary: Criar bloqueio de horario
      description: |
        Cria um novo bloqueio de horario para uma unidade.

        **Tipos de bloqueio:**
        - Dia inteiro: omitir `time_start` e `time_end`
        - Horario especifico: informar `time_start` e `time_end`

        Requer autenticacao com role `recrutamento` ou `admin`.
      operationId: createScheduleBlock
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScheduleBlockRequest'
            examples:
              diaInteiro:
                summary: Bloqueio de dia inteiro
                value:
                  id_unidade: 123
                  date_start: "2025-01-20"
                  date_end: "2025-01-24"
                  reason: "Ferias do coordenador"
              horarioEspecifico:
                summary: Bloqueio de horario especifico
                value:
                  id_unidade: 123
                  date_start: "2025-02-01"
                  date_end: "2025-02-01"
                  time_start: "14:00:00"
                  time_end: "16:00:00"
                  reason: "Reuniao pedagogica"
      responses:
        '201':
          description: Bloqueio criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleBlockResponse'
              example:
                success: true
                message: "Bloqueio criado com sucesso"
                block:
                  id_schedule_block: 1
                  id_unidade: 123
                  date_start: "2025-01-20"
                  date_end: "2025-01-24"
                  time_start: null
                  time_end: null
                  reason: "Ferias do coordenador"
                  created_at: "2025-01-10T10:00:00-03:00"
        '400':
          description: Dados invalidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                dataInvalida:
                  summary: Data fim anterior a data inicio
                  value:
                    success: false
                    error: "Data de fim deve ser igual ou posterior a data de inicio"
                horarioIncompleto:
                  summary: Horario de inicio sem horario de fim
                  value:
                    success: false
                    error: "Se time_start for informado, time_end tambem deve ser"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT obtido atraves do endpoint `/auth/login`.

        Enviar no header: `Authorization: Bearer {token}`
  responses:
    Unauthorized:
      description: Token nao fornecido ou invalido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Token nao fornecido ou invalido"
...
    InternalServerError:
      description: Erro interno do servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Erro interno do servidor"
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Mensagem de erro
          example: "Descricao do erro"
        errors:
          type: array
          description: Lista de erros de validacao
          items:
            type: object
            properties:
              msg:
                type: string
                example: "Campo obrigatorio"
              param:
                type: string
                example: "id_unidade"
              location:
                type: string
                example: "query"
        details:
          type: string
          description: Detalhes adicionais do erro (desenvolvimento)

    # --------------------------------------------------------------------------
    # Availability Schemas
    # --------------------------------------------------------------------------
    AvailabilityResponse:
      type: object
      required:
        - success
        - config
        - pagination
        - slots
      properties:
        success:
          type: boolean
          example: true
        config:
          $ref: '#/components/schemas/AvailabilityConfig'
        pagination:
          $ref: '#/components/schemas/WeekPagination'
        slots:
          type: array
          items:
            $ref: '#/components/schemas/Slot'
...
    BookingStatus:
      type: string
      enum:
        - agendado
        - compareceu
        - faltou
        - cancelado
      description: |
        Status do agendamento:
        - `agendado`: Agendamento confirmado, aguardando aula teste
        - `compareceu`: Candidato compareceu a aula teste
        - `faltou`: Candidato nao compareceu (no-show)
        - `cancelado`: Agendamento foi cancelado
...